Wrap ToC update logic into composable
